+++ src/base/gfx/colorgradient.cpp
@@ -4,31 +4,28 @@
 #include <stdexcept>
 #include <string>
 
-#include "base/math/segmentedfunc.h"
-#include "base/math/segmentedfunc.tpp"
 #include "base/text/smartstring.h"
 
 #include "color.h"
-
-template struct Math::SegmentedFunction<Gfx::Color,
-    Gfx::ColorGradient>;
 
 namespace Gfx
 {
 
 ColorGradient::ColorGradient(const std::string &stoplist)
 {
-	auto stopStrings = Text::SmartString::split(stoplist, ',', true);
-	auto pos = 0.0;
-	for (auto &stopString : stopStrings) {
-		auto parts = Text::SmartString::split(stopString, ' ', true);
-		if (parts.size() == 2) {
-			pos = std::stod(parts[1]);
-			stops.emplace_back(pos, Color(parts[0]));
-		}
-		else
-			throw std::logic_error(
-			    "invalid gradient stop: " + stopString);
+	for (const auto &stopString :
+	    Text::SmartString::split(stoplist, ',', true)) {
+		if (auto &&parts =
+		        Text::SmartString::split(stopString, ' ', true);
+		    parts.size() == 2)
+			if (auto pos = std::stod(parts[1]);
+			    std::isfinite(pos)
+			    && (stops.empty() || pos >= stops.back().pos)) {
+				stops.emplace_back(pos, Color::fromString(parts[0]));
+				continue;
+			}
+		throw std::logic_error(
+		    "invalid gradient stop: " + stopString);
 	}
 }