+++ src/base/math/segmentedfunc.h
@@ -3,6 +3,9 @@
 
 #include <vector>
 
+#include "interpolation.h"
+#include "range.h"
+
 namespace Math
 {
 
@@ -65,7 +68,22 @@ template <typename T, class CRTP> struct SegmentedFunction
 	[[nodiscard]] bool operator==(
 	    const SegmentedFunction &other) const = default;
 
-	[[nodiscard]] T operator()(double pos) const;
+	[[nodiscard]] T operator()(double pos) const
+	{
+		if (stops.empty()) return T();
+		if (stops.size() == 1 || pos < stops.front().pos)
+			return stops.front().value;
+		if (pos > stops.back().pos) return stops.back().value;
+
+		auto it = std::ranges::adjacent_find(stops,
+		    [pos](auto &&cur, auto &&next)
+		    {
+			    return cur.pos <= pos && pos <= next.pos;
+		    });
+		return interpolate(it->value,
+		    std::next(it)->value,
+		    Range{it->pos, std::next(it)->pos}.rescale(pos));
+	}
 };
 
 }