+++ test/unit/base/gfx/color.cpp
@@ -11,7 +11,7 @@ const static auto tests =
         .add_case("Convertable from hex string",
             []
             {
-	            const Gfx::Color color("#12345678");
+	            auto color = Gfx::Color::fromString("#12345678");
 	            check() << color.red == 0x12 / 255.0;
 	            check() << color.green == 0x34 / 255.0;
 	            check() << color.blue == 0x56 / 255.0;